Restrict endpoints to the proper roles

+++ TrappyKeepy.Api/Controllers/UserController.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using TrappyKeepy.Domain.Interfaces;
 using TrappyKeepy.Domain.Models;
-
+using TrappyKeepy.Service;
 
 namespace TrappyKeepy.Api.Controllers
 {
@@ -13,6 +13,8 @@ namespace TrappyKeepy.Api.Controllers
     public class UserController : ControllerBase
     {
         private readonly IUserService userService;
+        private readonly JwtManager jwtManager = new JwtManager();
+        private readonly Helpers help = new Helpers();
 
         public UserController(IUserService userService)
         {
@@ -24,9 +26,38 @@ namespace TrappyKeepy.Api.Controllers
         {
             try
             {
-                var serviceRequest = new UserServiceRequest(userDto);
-                var serviceResponse = await userService.Create(serviceRequest);
                 var response = new ControllerResponse();
+
+                // Verify the requester is authorized.
+                var authorized = jwtManager.DecodeJwt(help.ParseToken(Request.Headers));
+                if (authorized.type is not JwtType.ACCESS || authorized.role < UserRole.ADMIN)
+                {
+                    response.Fail("Unauthorized. Access denied.");
+                    return StatusCode(401, response);
+                }
+
+                if (userDto.Name is null || userDto.Password is null || userDto.Email is null)
+                {
+                    response.Fail("Name, password, and email are required to create a user.");
+                    return BadRequest(response);
+                }
+
+                // Prepare a user from the userDto to pass to the service.
+                var user = new User()
+                {
+                    Name = userDto.Name,
+                    Password = userDto.Password,
+                    Email = userDto.Email
+                };
+                if (userDto.Role is not null) user.Role = (short)userDto.Role;
+
+                // Prepare the service request.
+                var serviceRequest = new UserServiceRequest(user);
+
+                // Wait for the service response.
+                var serviceResponse = await userService.Create(serviceRequest);
+
+                // Send the controller response back to the client.
                 switch (serviceResponse.Outcome)
                 {
                     case OutcomeType.Error:
@@ -42,9 +73,9 @@ namespace TrappyKeepy.Api.Controllers
             }
             catch (Exception)
             {
-                // TODO: Log exception somewhere?
                 return StatusCode(500);
             }
+
             // Default to error if unknown outcome from the service.
             return StatusCode(500);
         }
@@ -54,9 +85,23 @@ namespace TrappyKeepy.Api.Controllers
         {
             try
             {
-                var serviceRequest = new UserServiceRequest();
-                var serviceResponse = await userService.ReadAll(serviceRequest);
                 var response = new ControllerResponse();
+
+                // Verify the requester is authorized.
+                var authorized = jwtManager.DecodeJwt(help.ParseToken(Request.Headers));
+                if (authorized.type is not JwtType.ACCESS || authorized.role < UserRole.ADMIN)
+                {
+                    response.Fail("Unauthorized. Access denied.");
+                    return StatusCode(401, response);
+                }
+
+                // Prepare the service request.
+                var serviceRequest = new UserServiceRequest();
+
+                // Wait for the service response.
+                var serviceResponse = await userService.ReadAll(serviceRequest);
+
+                // Send the controller response back to the client.
                 switch (serviceResponse.Outcome)
                 {
                     case OutcomeType.Error:
@@ -72,9 +117,9 @@ namespace TrappyKeepy.Api.Controllers
             }
             catch (Exception)
             {
-                // TODO: Log exception somewhere?
                 return StatusCode(500);
             }
+
             // Default to error if unknown outcome from the service.
             return StatusCode(500);
         }
@@ -84,9 +129,23 @@ namespace TrappyKeepy.Api.Controllers
         {
             try
             {
-                var serviceRequest = new UserServiceRequest(id);
-                var serviceResponse = await userService.ReadById(serviceRequest);
                 var response = new ControllerResponse();
+
+                // Verify the requester is authorized.
+                var authorized = jwtManager.DecodeJwt(help.ParseToken(Request.Headers));
+                if (authorized.type is not JwtType.ACCESS || authorized.role < UserRole.ADMIN)
+                {
+                    response.Fail("Unauthorized. Access denied.");
+                    return StatusCode(401, response);
+                }
+
+                // Prepare the service request.
+                var serviceRequest = new UserServiceRequest(id);
+
+                // Wait for the service response.
+                var serviceResponse = await userService.ReadById(serviceRequest);
+
+                // Send the controller response back to the client.
                 switch (serviceResponse.Outcome)
                 {
                     case OutcomeType.Error:
@@ -102,22 +161,58 @@ namespace TrappyKeepy.Api.Controllers
             }
             catch (Exception)
             {
-                // TODO: Log exception somewhere?
                 return StatusCode(500);
             }
+
             // Default to error if unknown outcome from the service.
             return StatusCode(500);
         }
 
-        // TODO: UpdateById
         [HttpPut("")]
         public async Task<ActionResult> UpdateById([FromBody] UserDto userDto)
         {
             try
             {
-                var serviceRequest = new UserServiceRequest(userDto);
-                var serviceResponse = await userService.UpdateById(serviceRequest);
                 var response = new ControllerResponse();
+
+                // Verify the requester is authorized.
+                var authorized = jwtManager.DecodeJwt(help.ParseToken(Request.Headers));
+                if (authorized.type is not JwtType.ACCESS || authorized.role < UserRole.ADMIN)
+                {
+                    response.Fail("Unauthorized. Access denied.");
+                    return StatusCode(401, response);
+                }
+
+                if (userDto.Id is null || userDto.Id == Guid.Empty || (Guid)userDto.Id == Guid.Empty)
+                {
+                    response.Fail("User id is required to update a user.");
+                    return BadRequest(response);
+                }
+
+                // Prepare a user from the userDto to pass to the service.
+                var user = new User() { Id = (Guid)userDto.Id };
+                if (userDto.Name is not null) user.Name = userDto.Name;
+                if (userDto.Email is not null) user.Email = userDto.Email;
+
+                // Determine if we are updating the user role.
+                if (userDto.Role is not null)
+                {
+                    user.Role = (short)userDto.Role;
+                }
+                else
+                {
+                    // If we aren't updating the user role, set it to -1.
+                    // The repository will only update role if role is >= 0.
+                    user.Role = -1;
+                }
+
+                // Prepare the service request.
+                var serviceRequest = new UserServiceRequest(user);
+
+                // Wait for the service response.
+                var serviceResponse = await userService.UpdateById(serviceRequest);
+
+                // Send the controller response back to the client.
                 switch (serviceResponse.Outcome)
                 {
                     case OutcomeType.Error:
@@ -133,9 +228,9 @@ namespace TrappyKeepy.Api.Controllers
             }
             catch (Exception)
             {
-                // TODO: Log exception somewhere?
                 return StatusCode(500);
             }
+
             // Default to error if unknown outcome from the service.
             return StatusCode(500);
         }
@@ -145,9 +240,36 @@ namespace TrappyKeepy.Api.Controllers
         {
             try
             {
-                var serviceRequest = new UserServiceRequest(userDto);
-                var serviceResponse = await userService.UpdatePasswordById(serviceRequest);
                 var response = new ControllerResponse();
+
+                // Verify the requester is authorized.
+                var authorized = jwtManager.DecodeJwt(help.ParseToken(Request.Headers));
+                if (authorized.type is not JwtType.ACCESS || authorized.role < UserRole.ADMIN)
+                {
+                    response.Fail("Unauthorized. Access denied.");
+                    return StatusCode(401, response);
+                }
+
+                if (userDto.Id is null || userDto.Id == Guid.Empty || (Guid)userDto.Id == Guid.Empty || userDto.Password is null)
+                {
+                    response.Fail("User id and password are required to update a user pasword.");
+                    return BadRequest(response);
+                }
+
+                // Prepare a user from the userDto to pass to the service.
+                var user = new User()
+                {
+                    Id = (Guid)userDto.Id,
+                    Password = userDto.Password
+                };
+
+                // Prepare the service request.
+                var serviceRequest = new UserServiceRequest(user);
+
+                // Wait for the service response.
+                var serviceResponse = await userService.UpdatePasswordById(serviceRequest);
+
+                // Send the controller response back to the client.
                 switch (serviceResponse.Outcome)
                 {
                     case OutcomeType.Error:
@@ -163,9 +285,9 @@ namespace TrappyKeepy.Api.Controllers
             }
             catch (Exception)
             {
-                // TODO: Log exception somewhere?
                 return StatusCode(500);
             }
+
             // Default to error if unknown outcome from the service.
             return StatusCode(500);
         }
@@ -175,9 +297,29 @@ namespace TrappyKeepy.Api.Controllers
         {
             try
             {
-                var serviceRequest = new UserServiceRequest(id);
-                var serviceResponse = await userService.DeleteById(serviceRequest);
                 var response = new ControllerResponse();
+
+                // Verify the requester is authorized.
+                var authorized = jwtManager.DecodeJwt(help.ParseToken(Request.Headers));
+                if (authorized.type is not JwtType.ACCESS || authorized.role < UserRole.ADMIN)
+                {
+                    response.Fail("Unauthorized. Access denied.");
+                    return StatusCode(401, response);
+                }
+
+                if (id == Guid.Empty || (Guid)id == Guid.Empty)
+                {
+                    response.Fail("User id is required to delete a user.");
+                    return BadRequest(response);
+                }
+
+                // Prepare the service request.
+                var serviceRequest = new UserServiceRequest(id);
+
+                // Wait for the service response.
+                var serviceResponse = await userService.DeleteById(serviceRequest);
+
+                // Send the controller response back to the client.
                 switch (serviceResponse.Outcome)
                 {
                     case OutcomeType.Error:
@@ -193,9 +335,9 @@ namespace TrappyKeepy.Api.Controllers
             }
             catch (Exception)
             {
-                // TODO: Log exception somewhere?
                 return StatusCode(500);
             }
+
             // Default to error if unknown outcome from the service.
             return StatusCode(500);
         }
@@ -206,9 +348,28 @@ namespace TrappyKeepy.Api.Controllers
         {
             try
             {
-                var serviceRequest = new UserServiceRequest(userDto);
-                var serviceResponse = await userService.Authenticate(serviceRequest);
                 var response = new ControllerResponse();
+
+                if (userDto.Email is null || userDto.Password is null)
+                {
+                    response.Fail("User email and password are required to authenticate a user.");
+                    return BadRequest(response);
+                }
+
+                // Prepare a user from the userDto to pass to the service.
+                var user = new User()
+                {
+                    Email = userDto.Email,
+                    Password = userDto.Password
+                };
+
+                // Prepare the service request.
+                var serviceRequest = new UserServiceRequest(user);
+
+                // Wait for the service response.
+                var serviceResponse = await userService.Authenticate(serviceRequest);
+
+                // Send the controller response back to the client.
                 switch (serviceResponse.Outcome)
                 {
                     case OutcomeType.Error:
@@ -224,9 +385,9 @@ namespace TrappyKeepy.Api.Controllers
             }
             catch (Exception)
             {
-                // TODO: Log exception somewhere?
                 return StatusCode(500);
             }
+
             // Default to error if unknown outcome from the service.
             return StatusCode(500);
         }
